requirements.txt
selenium==3.141.0
python_telegram_bot==12.5.1
